tests/wpt/tests/svg/animations/invalid-attribute-type-does-not-block-begin-event.html

<!DOCTYPE html>
<title>beginEvent must fire for a delayed animation when an invalid attribute type group exists</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<svg>
  <!-- Invalid animation: attributeType="CSS" on a non-CSS attribute. -->
  <rect>
    <animate attributeName="href" attributeType="CSS" values="foo"/>
  </rect>

  <!-- Valid animation with a delayed begin. -->
  <rect width="100" height="100" fill="red">
    <animate id="valid" attributeName="fill" values="green" begin="0.05s"/>
  </rect>
</svg>
<script>
async_test(t => {
  document.getElementById("valid").addEventListener("beginEvent", t.step_func_done(() => {
    assert_true(true, "beginEvent fired for delayed animation");
  }));
});
</script>
